Add current buddypress.svn.wordpress.org/trunk (2.1) .pot file for translation.
At time of writing, there exists a .pot in this repo's `bp-languages` folder, but when we update this repo's trunk for future BuddyPress releases (beta, RCS, and finals), the .pot file will have been moved to the root of trunk (the `bp-langauges` folder has been removed).
We're adding this .pot file early so we can ask for updates to translate.wordpress.org/projects/buddypress/ to accommodate this.

See background for https://buddypress.trac.wordpress.org/ticket/5621


git-svn-id: https://plugins.svn.wordpress.org/buddypress/trunk@965281 b8457f37-d9ea-0310-8a92-e5e31aec5664